Fixed the rest of the errors, renamed some calls for simplification and uniformity, updated exception handling to use ServiceException. Seriously updated handlers, Passes all but one test in StandardAPITests

+++ server/src/main/java/dataAccess/MemoryAuthDAO.java
@@ -3,35 +3,43 @@ package dataAccess;
 import model.AuthData;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.UUID;
 
 public class MemoryAuthDAO implements AuthDAO {
-    private final Map<String, AuthData> authDatabase;
+    private final Map<String, AuthData> authDatabase = new HashMap<>();
 
-    public MemoryAuthDAO() {
-        authDatabase = new HashMap<>();
+    @Override
+    public AuthData createAuth(String username) throws DataAccessException {
+        String authToken = UUID.randomUUID().toString();
+        AuthData authData = new AuthData(username, authToken);
+        authDatabase.put(authToken, authData);
+        return authData;
     }
 
     @Override
-    public void addAuth(AuthData authData) {
+    public void addAuth(AuthData authData) throws DataAccessException {
         authDatabase.put(authData.authToken(), authData);
     }
 
     @Override
-    public void deleteAuth(String authToken) {
+    public void deleteAuth(String authToken) throws DataAccessException {
+        if (!authDatabase.containsKey(authToken)) {
+            throw new DataAccessException("Error: unauthorized");
+        }
         authDatabase.remove(authToken);
     }
 
     @Override
-    public AuthData getAuth(String authToken) throws UnauthorizedException {
+    public AuthData getAuth(String authToken) throws DataAccessException {
         AuthData authData = authDatabase.get(authToken);
         if (authData == null) {
-            throw new UnauthorizedException("Error: unauthorized");
+            throw new DataAccessException("Error: unauthorized");
         }
         return authData;
     }
 
     @Override
-    public void clear() {
+    public void clear() throws DataAccessException {
         authDatabase.clear();
     }
 }
